+++ api/agents/utils.py
@@ -4,6 +4,20 @@ import json
 from typing import Any, Dict
 
 
+class BaseAgent:
+    """Base class for agents."""
+
+    def __init__(self, queries_history: list, result_history: list):
+        """Initialize the agent with query and result history."""
+        if result_history is None:
+            self.messages = []
+        else:
+            self.messages = []
+            for query, result in zip(queries_history[:-1], result_history):
+                self.messages.append({"role": "user", "content": query})
+                self.messages.append({"role": "assistant", "content": result})
+
+
 def parse_response(response: str) -> Dict[str, Any]:
     """
     Parse Claude's response to extract the analysis.
